Extract NLG annotation Name resolution
