+++ jl4-core/jl4-core.cabal
@@ -74,6 +74,7 @@ library
     L4.Syntax
     L4.Transform
     L4.TypeCheck
+    L4.TypeCheck.Annotation
     L4.TypeCheck.Environment
     L4.TypeCheck.Types
     L4.TypeCheck.With
