+++ jl4-core/src/L4/TypeCheck.hs
@@ -77,10 +77,10 @@ import qualified Base.Map as Map
 import qualified Base.Set as Set
 import qualified Base.Text as Text
 import L4.Annotation
-import L4.Lexer (PosToken(..))
 import L4.Parser.SrcSpan (prettySrcRange)
 import L4.Print (prettyLayout, quotedName)
 import L4.Syntax
+import L4.TypeCheck.Annotation
 import L4.TypeCheck.Environment as X
 import L4.TypeCheck.Types as X
 import L4.TypeCheck.With as X
@@ -164,12 +164,6 @@ runCheckUnique c s =
     [(w, s')] -> (w, s')
     _ -> error "internal error: expected unique result, got several"
 
-choose :: [Check a] -> Check a
-choose = asum
-
-anyOf :: [a] -> Check a
-anyOf = asum . fmap pure
-
 -- | Biased choice. Only takes the second option if the first fails.
 --
 orElse :: Check a -> Check a -> Check a
@@ -243,28 +237,6 @@ softprune m = do
     in
       proc candidates
 
-step :: Check Int
-step = do
-  current <- use #supply
-  let next = current + 1
-  assign #supply next
-  pure current
-
-newUnique :: Check Unique
-newUnique = do
-  i <- step
-  pure (MkUnique 'c' i)
-
-fresh :: RawName -> Check (Type' Resolved)
-fresh prefix = do
-  i <- step
-  pure (InfVar emptyAnno prefix i)
-
-addError :: CheckError -> Check ()
-addError e = do
-  ctx <- use #errorContext
-  with (MkCheckErrorWithContext e ctx)
-
 -- | Should never return 'Nothing' if our system is OK.
 getEntityInfo :: Resolved -> Check (Maybe CheckEntity)
 getEntityInfo r = do
@@ -274,67 +246,6 @@ getEntityInfo r = do
       addError (MissingEntityInfo r)
       pure Nothing
     Just (_n, ce) -> pure (Just ce)
-
-lookupRawNameInEnvironment :: RawName -> Check [(Unique, Name, CheckEntity)]
-lookupRawNameInEnvironment n = do
-  env <- use #environment
-  ei  <- use #entityInfo
-  let
-    proc :: Unique -> Maybe (Unique, Name, CheckEntity)
-    proc u = (\ (o, ce) -> (u, o, ce)) <$> Map.lookup u ei
-
-    candidates :: [(Unique, Name, CheckEntity)]
-    candidates = mapMaybe proc (Map.findWithDefault [] n env)
-
-  -- traceM ("Trying to look up " <> Text.unpack (prettyLayout n))
-  -- traceM ("Returning with " <> show (length candidates) <> " candidates")
-  -- traverse_ (traceM . debugCandidate) candidates
-  pure candidates
-
--- debugCandidate :: (Unique, Name, CheckEntity) -> String
--- debugCandidate (u, _, KnownType kind _)   = show u <> " of kind " <> show kind
--- debugCandidate (u, _, KnownTerm _ _)      = show u <> " is a term"
--- debugCandidate (u, _, KnownTypeVariable)  = show u <> " is a type variable"
-
-
-resolveTerm' :: (TermKind -> Bool) -> Name -> Check (Resolved, Type' Resolved)
-resolveTerm' p n = do
-  options <- lookupRawNameInEnvironment (rawName n)
-  case mapMaybe proc options of
-    [] -> do
-      v <- fresh (rawName n)
-      rn <- outOfScope (setAnnResolvedType v n) v
-      pure (rn, v)
-    [x] -> pure x
-    xs -> anyOf xs <|> do
-      v <- fresh (rawName n)
-      rn <- ambiguousTerm (setAnnResolvedType v n) xs
-      pure (rn, v)
-  where
-    proc :: (Unique, Name, CheckEntity) -> Maybe (Resolved, Type' Resolved)
-    proc (u, o, KnownTerm t tk) | p tk = Just (Ref (setAnnResolvedType t n) u o, t)
-    proc _                             = Nothing
-
-resolveTerm :: Name -> Check (Resolved, Type' Resolved)
-resolveTerm = resolveTerm' (const True)
-
-_resolveSelector :: Name -> Check (Resolved, Type' Resolved)
-_resolveSelector = resolveTerm' (== Selector)
-
-resolveConstructor :: Name -> Check (Resolved, Type' Resolved)
-resolveConstructor = resolveTerm' (== Constructor)
-
-setAnnResolvedType ::
-     (HasAnno a, AnnoToken a ~ PosToken, AnnoExtra a ~ Extension)
-  => Type' Resolved -> a -> a
-setAnnResolvedType t x =
-  setAnno (set annInfo (Just (TypeInfo t)) (getAnno x)) x
-
-setAnnResolvedKind ::
-     (HasAnno a, AnnoToken a ~ PosToken, AnnoExtra a ~ Extension)
-  => Kind -> a -> a
-setAnnResolvedKind k x =
-  setAnno (set annInfo (Just (KindInfo k)) (getAnno x)) x
 
 instantiate :: Type' Resolved -> Check (Type' Resolved)
 instantiate (Forall _ann ns t) = do
@@ -419,24 +330,6 @@ ref n a =
     (u, o) = getUniqueName a
   in
     pure (Ref n u o)
-
-outOfScope :: Name -> Type' Resolved -> Check Resolved
-outOfScope n t = do
-  addError (OutOfScopeError n t)
-  u <- newUnique
-  pure (OutOfScope u n)
-
-ambiguousTerm :: Name -> [(Resolved, Type' Resolved)] -> Check Resolved
-ambiguousTerm n xs = do
-  addError (AmbiguousTermError n xs)
-  u <- newUnique
-  pure (OutOfScope u n)
-
-ambiguousType :: Name -> [(Resolved, Kind)] -> Check Resolved
-ambiguousType n xs = do
-  addError (AmbiguousTypeError n xs)
-  u <- newUnique
-  pure (OutOfScope u n)
 
 inferDeclare :: Declare Name -> Check (Declare Resolved)
 inferDeclare (MkDeclare ann tysig appForm t) = do
@@ -574,205 +467,6 @@ inferDecide (MkDecide ann tysig appForm expr) = do
   extend
   pure rd
 
-nlgDecide :: Decide Resolved -> Check (Decide Resolved)
-nlgDecide (MkDecide ann tySig appForm body) =
-  MkDecide ann
-    <$> nlgTypeSig tySig
-    <*> nlgAppForm appForm
-    <*> nlgExpr body
-
-nlgExpr :: Expr Resolved -> Check (Expr Resolved)
-nlgExpr = \case
-    And ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ And ann e1' e2'
-    Or ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Or ann e1' e2'
-    Implies ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Implies ann e1' e2'
-    Equals ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Equals ann e1' e2'
-    Not ann e -> do
-      e' <- nlgExpr e
-      pure $ Not ann e'
-    Plus ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Plus ann e1' e2'
-    Minus ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Minus ann e1' e2'
-    Times ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Times ann e1' e2'
-    DividedBy ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ DividedBy ann e1' e2'
-    Modulo ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Modulo ann e1' e2'
-    Cons ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Cons ann e1' e2'
-    Leq ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Leq ann e1' e2'
-    Lt ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Lt ann e1' e2'
-    Gt ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Gt ann e1' e2'
-    Geq ann e1 e2 -> do
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ Geq ann e1' e2'
-    Proj ann e1 n -> do
-      e1' <- nlgExpr e1
-      n' <- resolveNlgAnnotationInResolved n
-      pure $ Proj ann e1' n'
-    Var ann v -> do
-      v' <- resolveNlgAnnotationInResolved v
-      pure $ Var ann v'
-    Lam ann sig body -> do
-      -- Since the parameters in the lambda bring new variables into
-      -- scope, we have to resolve the annotations when checking the 'Lam'
-      -- case. Thus, we don't need to traverse it here again.
-      pure $ Lam ann sig body
-    App ann n ns -> do
-      n' <- resolveNlgAnnotationInResolved n
-      ns' <- traverse nlgExpr ns
-      pure $ App ann n' ns'
-    AppNamed ann n ns order -> do
-      n' <- resolveNlgAnnotationInResolved n
-      ns' <- traverse nlgNamedExpr ns
-      pure $ AppNamed ann n' ns' order
-    IfThenElse ann b e1 e2 -> do
-      b' <- nlgExpr b
-      e1' <- nlgExpr e1
-      e2' <- nlgExpr e2
-      pure $ IfThenElse ann b' e1' e2'
-    Consider ann e branches  -> do
-      e' <- nlgExpr e
-      -- Since the bindings in the branches bring new variables into
-      -- scope, we have to resolve the annotations when checking the 'Consider'
-      -- case. Thus, we don't need to traverse it here again.
-      pure $ Consider ann e' branches
-    expr@Lit{} -> do
-      pure expr
-    List ann es -> do
-      es' <- traverse nlgExpr es
-      pure $ List ann es'
-    Where ann e lcl -> do
-      -- Since the bindings in the 'LocalDecl' bring new variables into
-      -- scope, we have to resolve the annotations when checking the 'Where'
-      -- case. Thus, we don't need to traverse it here again.
-      pure $ Where ann e lcl
-
-nlgPattern :: Pattern Resolved -> Check (Pattern Resolved)
-nlgPattern = \case
-  PatVar ann n ->
-    PatVar ann
-      <$> resolveNlgAnnotationInResolved n
-  PatApp ann n pats ->
-    PatApp ann
-      <$> resolveNlgAnnotationInResolved n
-      <*> traverse nlgPattern pats
-  PatCons ann pat pats ->
-    PatCons ann
-      <$> nlgPattern pat
-      <*> nlgPattern pats
-
-nlgLocalDecl :: LocalDecl Resolved -> Check (LocalDecl Resolved)
-nlgLocalDecl = \case
-  LocalDecide ann decide ->
-    LocalDecide ann
-      <$> nlgDecide decide
-  LocalAssume ann assume ->
-    LocalAssume ann
-      <$> nlgAssume assume
-
-nlgAssume :: Assume Resolved -> Check (Assume Resolved)
-nlgAssume = \case
-  MkAssume ann tySig appForm mTy ->
-    MkAssume ann
-      <$> nlgTypeSig tySig
-      <*> nlgAppForm appForm
-      <*> traverse nlgType mTy
-
-nlgNamedExpr :: NamedExpr Resolved -> Check (NamedExpr Resolved)
-nlgNamedExpr = \case
-  MkNamedExpr ann n expr ->
-    MkNamedExpr ann
-      <$> resolveNlgAnnotationInResolved n
-      <*> nlgExpr expr
-
-nlgAppForm :: AppForm Resolved -> Check (AppForm Resolved)
-nlgAppForm (MkAppForm ann n ns maka) =
-  MkAppForm ann
-    <$> resolveNlgAnnotationInResolved n
-    <*> traverse resolveNlgAnnotationInResolved ns
-    <*> traverse resolveNlgAnnotation maka
-
-nlgTypeSig :: TypeSig Resolved -> Check (TypeSig Resolved)
-nlgTypeSig (MkTypeSig ann givenSig mGivethSig) =
-  MkTypeSig ann
-    <$> nlgGivenSig givenSig
-    <*> traverse nlgGivethSig mGivethSig
-
-nlgGivethSig :: GivethSig Resolved -> Check (GivethSig Resolved)
-nlgGivethSig (MkGivethSig ann ty) =
-  MkGivethSig ann
-    <$> nlgType ty
-
-nlgType :: Type' Resolved -> Check (Type' Resolved)
-nlgType = \case
-  Type   ann ->
-    pure $ Type ann
-  TyApp  ann n tys ->
-    TyApp ann
-      <$> resolveNlgAnnotationInResolved n
-      <*> traverse nlgType tys
-  Fun    ann optNamedTypes ty ->
-    Fun ann
-      <$> traverse nlgOptionallyNamedType optNamedTypes
-      <*> nlgType ty
-  Forall ann ns ty ->
-    Forall ann <$> traverse resolveNlgAnnotationInResolved ns <*> nlgType ty
-  InfVar ann name uniq ->
-    pure $ InfVar ann name uniq
-
-nlgOptionallyNamedType :: OptionallyNamedType Resolved -> Check (OptionallyNamedType Resolved)
-nlgOptionallyNamedType (MkOptionallyNamedType ann mName ty) =
-  MkOptionallyNamedType ann
-    <$> traverse resolveNlgAnnotationInResolved mName
-    <*> nlgType ty
-
-nlgGivenSig :: GivenSig Resolved -> Check (GivenSig Resolved)
-nlgGivenSig (MkGivenSig ann ns) =
-  MkGivenSig ann
-    <$> traverse nlgOptionallyTypedName ns
-
-nlgOptionallyTypedName :: OptionallyTypedName Resolved -> Check (OptionallyTypedName Resolved)
-nlgOptionallyTypedName (MkOptionallyTypedName ann n mty) =
-  MkOptionallyTypedName ann
-    <$> resolveNlgAnnotationInResolved n
-    <*> traverse nlgType mty
 
 -- | We allow the following cases:
 --
@@ -1193,37 +887,6 @@ scope m = do
   assign #environment savedEnv
   assign #entityInfo savedEi
   pure a
-
-resolveNlgAnnotation ::
-  (HasAnno a, AnnoToken a ~ PosToken, AnnoExtra a ~ Extension) =>
-  a ->
-  Check a
-resolveNlgAnnotation a = do
-  case a ^. annoOf % annNlg of
-    Nothing -> pure a
-    Just nlgM -> do
-      resolvedNlg <- case nlgM of
-        MkInvalidNlg{} -> pure nlgM
-        MkResolvedNlg{} -> pure nlgM
-        MkParsedNlg ann frags -> do
-          resolvedFrags <- traverse resolveNlgFragment frags
-          pure $ MkResolvedNlg ann resolvedFrags
-      pure $ a & annoOf % annNlg ?~ resolvedNlg
-
-resolveNlgFragment :: NlgFragment Name -> Check (NlgFragment Resolved)
-resolveNlgFragment = \case
-  MkNlgText ann t -> pure $ MkNlgText ann t
-  MkNlgRef ann n ->
-    MkNlgRef ann . fst <$> resolveTerm n
-
-resolveNlgAnnotationInResolved :: Resolved -> Check Resolved
-resolveNlgAnnotationInResolved = \case
-  Def uniq name -> do
-    Def uniq <$> resolveNlgAnnotation name
-  Ref refName uniq origName -> do
-    Ref refName uniq <$> resolveNlgAnnotation origName
-  OutOfScope uniq origName -> do
-    OutOfScope uniq <$> resolveNlgAnnotation origName
 
 setErrorContext :: (CheckErrorContext -> CheckErrorContext) -> Check ()
 setErrorContext f =
